+++ src/display.cpp
@@ -88,8 +88,14 @@ void setProperties(NC::Menu<T> &menu, const MPD::Song &s, const SongList &list,
 		auto next = list.beginS() + drawn_pos + 1;
 		if (next != list.endS())
 		{
-			if (next->song() != nullptr && next->song()->getAlbum() != s.getAlbum())
-				separate_albums = true;
+			if (next->song() != nullptr)
+			{
+				// Draw a separator when the next album is different than the current
+				// one. In case there are two albums with the same name, but a different
+				// artist, compare also artists.
+				separate_albums = next->song()->getAlbum() != s.getAlbum()
+				               || next->song()->getArtist() != s.getArtist();
+			}
 		}
 	}
 	if (separate_albums)