+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix crash on startup with Visualizer as the initial screen.
 * Fix crash on startup with Browser as the initial screen.
 * Show the Visualizer immediately if it's the initial screen.
+* Draw a separator between albums with the same name, but a different artist.
 
 # ncmpcpp-0.9 (2020-12-20)
 * Fix various Mopidy specific bugs.